+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Always include `wp-admin/includes/plugin.php` so we have `is_active_plugin()` in the front-end.
+- Directory routing in cases for routes with `{directory}?{key}={value}`.
 - Added missing return when terms are empty in `LangInterface::get_terms_for_source`.
 - Bulk edit not working on non english back-offices and for non `post` post types.
 - No language filters not showing up for non `post` post types.